% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{Delta_z}
\alias{Delta_z}
\title{Variation of the thickness of soil layer}
\usage{
Delta_z(
  f_agg = 3,
  Delta_z_min,
  My_mic,
  Mo_mic,
  My_mes,
  Mo_mes,
  phi_mac,
  gamma_o
)
}
\arguments{
\item{f_agg}{an aggregation factor (m3 pore space m-3 organic matter) defined as the slope of the linear relationship assumed between the volume of aggregation pore space \eqn{V_{agg}}, and the volume of organic matter \eqn{V_{s_o}}}

\item{Delta_z_min}{minimal soil thickness if no organic matter was present}

\item{My_mic}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{Mo_mic}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{My_mes}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{Mo_mes}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{phi_mac}{macroporosity}

\item{gamma_o}{density of organic matter}
}
\value{
one single value
}
\description{
This function calculates the variation of the thickness of soil layer as a function of organic matter
}
