% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{f_text_mic}
\alias{f_text_mic}
\title{Proportion of micropores}
\usage{
f_text_mic(clay, phi_min)
}
\arguments{
\item{clay}{soil clay fraction}

\item{phi_min}{minimal porosity, user defined}
}
\value{
one single value
}
\description{
This function calculates the proportion of the textural pore space that comprises micropores. It is used in \code{\link{pore_frac}}
}
