% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{gamma_b}
\alias{gamma_b}
\title{Soil bulk density}
\usage{
gamma_b(
  My_mic,
  Mo_mic,
  My_mes,
  Mo_mes,
  Delta_z_min,
  phi_min,
  gamma_o,
  gamma_m,
  f_agg,
  phi_mac
)
}
\arguments{
\item{My_mic}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{Mo_mic}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{My_mes}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{Mo_mes}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{Delta_z_min}{minimal soil thickness if no organic matter was present (cm)}

\item{phi_min}{minimal porosity, (\eqn{\frac{cm^3 of water}{cm^3 of soil}})}

\item{f_agg}{an aggregation factor (m3 pore space m-3 organic matter) defined as the slope of the linear relationship assumed between the volume of aggregation pore space \eqn{V_{agg}}, and the volume of organic matter \eqn{V_{s_o}} (dimensionless)}

\item{phi_mac}{macroporosity, (\eqn{\frac{cm^3 of water}{cm^3 of soil}})}
}
\value{
one single value
}
\description{
This function calculates the soil bulk density. It relies on \code{\link{Msm}} to calculate the mineral mass.
}
\seealso{
\code{\link{Msm}}
}
