% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{mass_balance}
\alias{mass_balance}
\title{Mass balance check}
\usage{
mass_balance(results, Im, Ir, sim_length)
}
\arguments{
\item{results}{results as from the function  \code{\link{run_Porous}}}
}
\value{
one single value
}
\description{
This function calculates the mass balance between the C fluxes and stock change and the inputs. It is used mainly as a diagnostic for model development.
}
\seealso{
\code{\link{Msm}}
}
