% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{Msm}
\alias{Msm}
\title{Mineral mass}
\usage{
Msm(Delta_z_min, phi_min, gamma_m)
}
\arguments{
\item{Delta_z_min}{minimal soil thickness if no organic matter was present (cm)}
}
\value{
one single value
}
\description{
This function calculates the mineral mass
}
\seealso{
\code{\link{Delta_z_min}}, \code{\link{gamma_b}}, \code{\link{f_som}}
}
