% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{phi_mic}
\alias{phi_mic}
\title{Microporosity}
\usage{
phi_mic(
  My_mic,
  Mo_mic,
  My_mes,
  Mo_mes,
  gamma_o,
  f_agg = 3,
  clay,
  Delta_z_min,
  phi_min = 1,
  phi_mac
)
}
\arguments{
\item{My_mic}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{Mo_mic}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{My_mes}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{Mo_mes}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{gamma_o}{density of organic matter}

\item{f_agg}{an aggregation factor (m3 pore space m−3 organic matter) defined as the slope of the linear relationship assumed between the volume of aggregation pore space \eqn{V_{agg}}, and the volume of organic matter \eqn{V_{s_o}}}

\item{clay}{fraction of clay content}

\item{Delta_z_min}{minimal soil thickness if no organic matter was present}

\item{phi_min}{minimal porosity, user defined}

\item{phi_mac}{macroporosity}
}
\value{
one single value
}
\description{
This function calculates the microporosity \eqn{\phi_{mic}} based on the variation of organic matter in the soil. It is used in \code{\link{pore_frac}}
}
