% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{pore_frac}
\alias{pore_frac}
\title{The main accessory function of the model}
\usage{
pore_frac(
  phi_mac = 0.2,
  clay = 0.2,
  Delta_z_min = 20,
  gamma_o = 1.2,
  My_mic,
  Mo_mic,
  My_mes,
  Mo_mes
)
}
\arguments{
\item{phi_mac}{macroporosity}

\item{clay}{fraction of clay content}

\item{Delta_z_min}{minimal soil thickness if no organic matter was present}

\item{gamma_o}{density of organic matter}

\item{My_mic}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{Mo_mic}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{My_mes}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{Mo_mes}{One of the four model pools (they are all summed up in this function for calculating the total)}
}
\value{
two values, the proportion of input in the mesopore and micropore Y pools
}
\description{
This function  calculates the proportion of inputs in each of the two youg pools deending on the organic matter content
}
