% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POROUS_testing_minimal.R
\name{Porous_test_min}
\alias{Porous_test_min}
\title{Purely for testing purposes}
\usage{
Porous_test_min(ky, ko, e, Im, Ir, F_prot, proportion)
}
\arguments{
\item{ky}{decomposition constant of the Young pool \eqn{frac{1}{year}}}

\item{ko}{decomposition constant of the Old pool \eqn{frac{1}{year}}}

\item{e}{efficiency, which is the transfer term between the pools and corresponds to the term h in the ICBM model in Kätterer et al. (2001) (dimensionless)}

\item{F_prot}{protection provided by the micropore space (dimensionless)}

\item{proportion}{this is a linearization term to make the proportion of the inputs between micro- and mesopores constant. If NULL (or not specified, since default is NULL) then the model is running as nonlinear, as in the original paper. If specified (must be between 0 and 1) then the model is linearized adopting this value as fixed proportion of inputs from roots going into the mesopore space (and its reciprocal into the micropore)}

\item{kmix}{mixing rate \eqn{frac{1}{year}}}
}
\value{
two values, the proportion (between 0 and 1) of input in the mesopore and micropore Y pools
}
\description{
Purely for testing purposes
}
\references{
Meurer, Katharina Hildegard Elisabeth, Claire Chenu, Elsa Coucheney, Anke Marianne Herrmann, Thomas Keller, Thomas Kätterer, David Nimblad Svensson, and Nicholas Jarvis. “Modelling Dynamic Interactions between Soil Structure and the Storage and Turnover of Soil Organic Matter.” Biogeosciences 17, no. 20 (October 19, 2020): 5025–42. https://doi.org/10.5194/bg-17-5025-2020. \cr
Kätterer, Thomas, and Olof Andrén. “The ICBM Family of Analytically Solved Models of Soil Carbon, Nitrogen and Microbial Biomass Dynamics — Descriptions and Application Examples.” Ecological Modelling 136, no. 2–3 (January 2001): 191–207. https://doi.org/10.1016/S0304-3800(00)00420-8.
}
\seealso{
\code{\link{run_Porous}}
}
