% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POROUS.R
\name{run_Porous}
\alias{run_Porous}
\title{Run Porous **STILL TESTING; NOT YET WORKING, use the deSolve function instead**}
\usage{
run_Porous(
  ky,
  ko,
  kmix,
  e,
  Im,
  Ir,
  F_prot,
  phi_mac,
  phi_min,
  clay,
  Delta_z_min,
  gamma_o,
  gamma_m,
  proportion = NULL,
  f_text_mic = NULL,
  f_agg,
  init,
  sim_length,
  sim_steps
)
}
\arguments{
\item{ky}{decomposition constant of the Young pool \eqn{frac{1}{year}}}

\item{ko}{decomposition constant of the Old pool \eqn{frac{1}{year}}}

\item{kmix}{mixing rate \eqn{frac{1}{year}}}

\item{e}{efficiency, which is the transfer term between the pools and corresponds to the term h in the ICBM model in Kätterer et al. (2001) (dimensionless)}

\item{F_prot}{protection provided by the micropore space (dimensionless)}

\item{phi_mac}{macroporosity, (\eqn{\frac{cm^3 of water}{cm^3 of soil}})}

\item{phi_min}{minimal porosity, (\eqn{\frac{cm^3 of water}{cm^3 of soil}})}

\item{clay}{fraction of clay content (dimensionless, between 0 and 1)}

\item{Delta_z_min}{minimal soil thickness if no organic matter was present (cm)}

\item{proportion}{this is a linearization term to make the proportion of the inputs between micro- and mesopores constant. If NULL (or not specified, since default is NULL) then the model is running as nonlinear, as in the original paper. If specified (must be between 0 and 1) then the model is linearized adopting this value as fixed proportion of inputs from roots going into the mesopore space (and its reciprocal into the micropore)}

\item{f_agg}{an aggregation factor (m3 pore space m-3 organic matter) defined as the slope of the linear relationship assumed between the volume of aggregation pore space \eqn{V_{agg}}, and the volume of organic matter \eqn{V_{s_o}} (dimensionless)}

\item{sim_steps}{steps of the simulation (fraction of one year)}

\item{sim_lenght}{length of the simulation (years)}
}
\value{
a data frame with the simulation of C stocks for each of the four pools, soil respiration for each of the four pools, bulk density variation and SOC concentration.
}
\description{
This function is a wrapper for running the main model \code{\link{Porous}}, which is a wrapper for the dual porosity decomposition model (equations 1 to 6) described in Meurer et al. (2020) implemented in  `SoilR`.
It then feeds the simulated SOC stocks to the functions \code{\link{gamma_b}} and \code{\link{f_som}} to calculate the variation of bulk density and C concentrations
}
\seealso{
\code{\link{Porous}}, \code{\link{gamma_b}}, \code{\link{f_som}}
}
