% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{Delta_z}
\alias{Delta_z}
\title{Variation of the thickness of soil layer}
\usage{
Delta_z(
  f_agg = f_agg,
  Delta_z_min,
  My_mic,
  Mo_mic,
  My_mes,
  Mo_mes,
  phi_mac,
  gamma_o
)
}
\arguments{
\item{f_agg}{an aggregation factor (m3 pore space m-3 organic matter) defined as the slope of the linear relationship assumed between the volume of aggregation pore space \eqn{V_{agg}}, and the volume of organic matter \eqn{V_{s_o}} (dimensionless)}

\item{Delta_z_min}{minimal soil thickness if no organic matter was present (cm)}

\item{My_mic}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{Mo_mic}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{My_mes}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{Mo_mes}{One of the four model pools (they are all summed up in this function for calculating the total)}

\item{phi_mac}{macroporosity, (\eqn{\frac{cm^3 of water}{cm^3 of soil}})}
}
\value{
one single value
}
\description{
This function calculates the variation of the thickness of soil layer as a function of organic matter. The parameter \eqn{f_{agg}} should be estimated from data on the relationship between bulk density (or its inverse, the specific volume) and soil organic matter content (see eq. 19 and fig. 4 in Meurer et al., 2020; from this data and other studies, a good average value of fagg should be around 3, which is the default value)
}
